chore(template): replace textual includes with modules
